refactor: improved common component structure
